+++ src/block-kit/container/Modal.tsx
@@ -10,11 +10,12 @@ import { DistributedProps } from '../../utils'
 import { plainText } from '../composition/utils'
 import { Select } from '../elements/Select'
 import { Textarea } from '../input/Textarea'
+import { availableActionTypes } from '../layout/Actions'
 import { Divider } from '../layout/Divider'
 import { Header } from '../layout/Header'
 import { Image } from '../layout/Image'
 import { Input, knownInputs } from '../layout/Input'
-import { Section } from '../layout/Section'
+import { Section, availableSectionAccessoryTypes } from '../layout/Section'
 import { Video } from '../layout/Video'
 import {
   PrivateMetadataTransformer,
@@ -124,13 +125,19 @@ type ModalProps = DistributedProps<BasicModalProps | WorkflowStepModalProps>
 const ModalBlocks = generateBlocksContainer({
   name: 'Modal',
   availableBlockTypes: {
-    actions: generateActionsValidator(),
+    actions: generateActionsValidator(
+      [...availableActionTypes].filter((v) => v !== 'workflow_button'),
+    ),
     context: true,
     divider: true,
     header: true,
     image: true,
     input: true,
-    section: generateSectionValidator(),
+    section: generateSectionValidator(
+      [...availableSectionAccessoryTypes].filter(
+        (v) => v !== 'workflow_button',
+      ),
+    ),
     video: true,
   },
   aliases: {