+++ src/block-kit/elements/UsersSelect.ts
@@ -52,8 +52,8 @@ export type UsersSelectProps = InputComponentProps<
 
 /**
  * The interactive component or input component for
- * [the `users_select` block element](https://api.slack.com/reference/block-kit/block-elements#users_select) and
- * [the `multi_users_select` block element](https://api.slack.com/reference/block-kit/block-elements#users_multi_select).
+ * [the `users_select` element](https://api.slack.com/reference/block-kit/block-elements#users_select) and
+ * [the `multi_users_select` element](https://api.slack.com/reference/block-kit/block-elements#users_multi_select).
  *
  * Provide a selectable menu element from a list of Slack users visible to the
  * current user in the active workspace.